+++ Tests/Redbird/FormattingTests.swift
@@ -139,20 +139,4 @@ class FormattingTests: XCTestCase {
         let str = try! InitialFormatter().format(obj)
         XCTAssertEqual(str, "*2\r\n*3\r\n:1\r\n:2\r\n:3\r\n*2\r\n+Foo\r\n-Bar\r\n")
     }
-    
-    func testPerf_LargeArray() {
-        let subinput: [RespObject] = [
-            RespBulkString(content: "large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here, large text right here"),
-            try! RespInteger(content: "1234567"),
-            RespError(content: "ERR Something went mildly wrong"),
-            RespNullArray(),
-            try! RespSimpleString(content: "Jokes"),
-            RespNullBulkString()
-        ]
-        let content: [RespObject] = Array(1..<100).map { _ in RespArray(content: subinput) }
-        let input = RespArray(content: content)
-        measureBlock {
-            _ = try! InitialFormatter().format(input)
-        }
-    }
 }
